+++ src/webview/features/entity-editor/components/TaskFields.tsx
@@ -1,4 +1,4 @@
-import { formatShortDate } from "@common/datePresentation";
+import { formatShortDate } from "@common/dates";
 import { TaskStatus } from "@common/documents";
 import { validateTaskConstraints } from "@services/schedule/scheduleConstraintService";
 import "@webview/components/Form.scss";